\name{FMsmsnReg-package}
\alias{FMsmsnReg-package}
\docType{package}
\title{Linear Regression Models with Finite Mixtures of Skew Heavy-Tailed Errors}
\description{
This package contains a principal function that performs to estimate the parameters of a regression model considering an error that follows a finite mixture of Skew Heavy-Tailed Errors, using an analytically simple and efficient EM-type algorithm for iteratively computing maximum likelihood estimates of the parameters.
}
\details{
\tabular{ll}{
Package: \tab FMsmsnsReg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-03-30\cr
License: \tab GPL (>=2)\cr
             }
}

\author{Luis Benites Sanchez \email{lbenitesanchez@gmail.com}, Rocio Paola Maehara \email{rmaeharaa@gmail.com} and Victor Hugo Lachos \email{hlachos@ime.unicamp.br}}

\references{
Basso, R. M., Lachos, V. H., Cabral, C. R., Ghosh, P., 2010. Robust mixture modeling based on scale mixtures of skew - normal distributions. \emph{Computational Statistics & Data Analysis}.

Lachos, V. H., Ghosh, P., Arellano-Valle, R. B., 2010. Likelihood based inference for skew-normal independent linear mixed models. \emph{Statistica Sinica} 20, 303 - 322.
}

\keyword{package}

\seealso{
\code{\link{FMsmsnReg}}
}

\examples{
#See examples for the FMsmsnReg function linked above.
}
